refactor: use scroll
